Alerting: Move alerting RTKQ client to api-clients package (#114546)
