--- packages/grafana-alerting/src/grafana/api/notifications/v0alpha1/const.ts
@@ -1,2 +1,0 @@
-export const VERSION = 'v0alpha1' as const;
-export const GROUP = 'notifications.alerting.grafana.app' as const;
