+++ packages/grafana-alerting/src/grafana/notificationPolicies/hooks/useMatchPolicies.test.ts
@@ -1,6 +1,6 @@
-import { VERSION } from '../../api/notifications/v0alpha1/const';
+import { API_VERSION, RoutingTree } from '@grafana/api-clients/rtkq/notifications.alerting/v0alpha1';
+
 import { LabelMatcherFactory, RouteFactory } from '../../api/notifications/v0alpha1/mocks/fakes/Routes';
-import { RoutingTree } from '../../api/notifications/v0alpha1/notifications.api.gen';
 import { Label } from '../../matchers/types';
 
 import { matchInstancesToRouteTrees } from './useMatchPolicies';
@@ -16,7 +16,7 @@ describe('matchInstancesToRouteTrees', () => {
     const trees: RoutingTree[] = [
       {
         kind: 'RoutingTree',
-        apiVersion: VERSION,
+        apiVersion: API_VERSION,
         metadata: { name: treeName },
         spec: {
           defaults: {
@@ -24,7 +24,6 @@ describe('matchInstancesToRouteTrees', () => {
           },
           routes: [route],
         },
-        status: {},
       },
     ];
 
@@ -51,7 +50,7 @@ describe('matchInstancesToRouteTrees', () => {
     const trees: RoutingTree[] = [
       {
         kind: 'RoutingTree',
-        apiVersion: VERSION,
+        apiVersion: API_VERSION,
         metadata: { name: treeName },
         spec: {
           defaults: {
@@ -59,7 +58,6 @@ describe('matchInstancesToRouteTrees', () => {
           },
           routes: [route],
         },
-        status: {},
       },
     ];
