+++ packages/grafana-alerting/tests/test-utils.tsx
@@ -13,7 +13,14 @@ import '@testing-library/jest-dom';
  * method which wraps the passed element in all of the necessary Providers,
  * so it can render correctly in the context of the application
  */
-const customRender = (ui: React.ReactNode, renderOptions: RenderOptions = {}) => {
+const customRender = (
+  ui: React.ReactNode,
+  renderOptions: RenderOptions = {}
+): {
+  renderResult: ReturnType<typeof render>;
+  user: ReturnType<typeof userEvent.setup>;
+  store: typeof store;
+} => {
   const user = userEvent.setup();
   const Providers = renderOptions.wrapper || getDefaultWrapper();
