+++ packages/grafana-api-clients/package.json
@@ -116,6 +116,14 @@
       "import": "./dist/esm/clients/rtkq/shorturl/v1beta1/index.mjs",
       "require": "./dist/cjs/clients/rtkq/shorturl/v1beta1/index.cjs"
     },
+    "./rtkq/notifications.alerting/v0alpha1": {
+      "import": "./src/clients/rtkq/notifications.alerting/v0alpha1/index.ts",
+      "require": "./src/clients/rtkq/notifications.alerting/v0alpha1/index.ts"
+    },
+    "./rtkq/rules.alerting/v0alpha1": {
+      "import": "./src/clients/rtkq/rules.alerting/v0alpha1/index.ts",
+      "require": "./src/clients/rtkq/rules.alerting/v0alpha1/index.ts"
+    },
     "./rtkq/historian.alerting/v0alpha1": {
       "@grafana-app/source": "./src/clients/rtkq/historian.alerting/v0alpha1/index.ts",
       "types": "./dist/types/clients/rtkq/historian.alerting/v0alpha1/index.d.ts",
