+++ packages/grafana-api-clients/src/clients/rtkq/notifications.alerting/v0alpha1/endpoints.gen.ts
@@ -1,5 +1,5 @@
-import { api } from './api';
-export const addTagTypes = ['API Discovery', 'AlertRule', 'RecordingRule'] as const;
+import { api } from './baseAPI';
+export const addTagTypes = ['API Discovery', 'Receiver', 'RoutingTree', 'TemplateGroup', 'TimeInterval'] as const;
 const injectedRtkApi = api
   .enhanceEndpoints({
     addTagTypes,
@@ -7,12 +7,12 @@ const injectedRtkApi = api
   .injectEndpoints({
     endpoints: (build) => ({
       getApiResources: build.query<GetApiResourcesApiResponse, GetApiResourcesApiArg>({
-        query: () => ({ url: `/apis/rules.alerting.grafana.app/v0alpha1/` }),
+        query: () => ({ url: `/` }),
         providesTags: ['API Discovery'],
       }),
-      listAlertRule: build.query<ListAlertRuleApiResponse, ListAlertRuleApiArg>({
+      listReceiver: build.query<ListReceiverApiResponse, ListReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules`,
+          url: `/receivers`,
           params: {
             pretty: queryArg.pretty,
             allowWatchBookmarks: queryArg.allowWatchBookmarks,
@@ -27,13 +27,13 @@ const injectedRtkApi = api
             watch: queryArg.watch,
           },
         }),
-        providesTags: ['AlertRule'],
+        providesTags: ['Receiver'],
       }),
-      createAlertRule: build.mutation<CreateAlertRuleApiResponse, CreateAlertRuleApiArg>({
+      createReceiver: build.mutation<CreateReceiverApiResponse, CreateReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules`,
+          url: `/receivers`,
           method: 'POST',
-          body: queryArg.alertRule,
+          body: queryArg.receiver,
           params: {
             pretty: queryArg.pretty,
             dryRun: queryArg.dryRun,
@@ -41,11 +41,11 @@ const injectedRtkApi = api
             fieldValidation: queryArg.fieldValidation,
           },
         }),
-        invalidatesTags: ['AlertRule'],
+        invalidatesTags: ['Receiver'],
       }),
-      deletecollectionAlertRule: build.mutation<DeletecollectionAlertRuleApiResponse, DeletecollectionAlertRuleApiArg>({
+      deletecollectionReceiver: build.mutation<DeletecollectionReceiverApiResponse, DeletecollectionReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules`,
+          url: `/receivers`,
           method: 'DELETE',
           params: {
             pretty: queryArg.pretty,
@@ -64,22 +64,22 @@ const injectedRtkApi = api
             timeoutSeconds: queryArg.timeoutSeconds,
           },
         }),
-        invalidatesTags: ['AlertRule'],
+        invalidatesTags: ['Receiver'],
       }),
-      getAlertRule: build.query<GetAlertRuleApiResponse, GetAlertRuleApiArg>({
+      getReceiver: build.query<GetReceiverApiResponse, GetReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}`,
+          url: `/receivers/${queryArg.name}`,
           params: {
             pretty: queryArg.pretty,
           },
         }),
-        providesTags: ['AlertRule'],
+        providesTags: ['Receiver'],
       }),
-      replaceAlertRule: build.mutation<ReplaceAlertRuleApiResponse, ReplaceAlertRuleApiArg>({
+      replaceReceiver: build.mutation<ReplaceReceiverApiResponse, ReplaceReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}`,
+          url: `/receivers/${queryArg.name}`,
           method: 'PUT',
-          body: queryArg.alertRule,
+          body: queryArg.receiver,
           params: {
             pretty: queryArg.pretty,
             dryRun: queryArg.dryRun,
@@ -87,11 +87,11 @@ const injectedRtkApi = api
             fieldValidation: queryArg.fieldValidation,
           },
         }),
-        invalidatesTags: ['AlertRule'],
+        invalidatesTags: ['Receiver'],
       }),
-      deleteAlertRule: build.mutation<DeleteAlertRuleApiResponse, DeleteAlertRuleApiArg>({
+      deleteReceiver: build.mutation<DeleteReceiverApiResponse, DeleteReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}`,
+          url: `/receivers/${queryArg.name}`,
           method: 'DELETE',
           params: {
             pretty: queryArg.pretty,
@@ -102,11 +102,11 @@ const injectedRtkApi = api
             propagationPolicy: queryArg.propagationPolicy,
           },
         }),
-        invalidatesTags: ['AlertRule'],
+        invalidatesTags: ['Receiver'],
       }),
-      updateAlertRule: build.mutation<UpdateAlertRuleApiResponse, UpdateAlertRuleApiArg>({
+      updateReceiver: build.mutation<UpdateReceiverApiResponse, UpdateReceiverApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}`,
+          url: `/receivers/${queryArg.name}`,
           method: 'PATCH',
           body: queryArg.patch,
           params: {
@@ -117,49 +117,11 @@ const injectedRtkApi = api
             force: queryArg.force,
           },
         }),
-        invalidatesTags: ['AlertRule'],
+        invalidatesTags: ['Receiver'],
       }),
-      getAlertRuleStatus: build.query<GetAlertRuleStatusApiResponse, GetAlertRuleStatusApiArg>({
+      listRoutingTree: build.query<ListRoutingTreeApiResponse, ListRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}/status`,
-          params: {
-            pretty: queryArg.pretty,
-          },
-        }),
-        providesTags: ['AlertRule'],
-      }),
-      replaceAlertRuleStatus: build.mutation<ReplaceAlertRuleStatusApiResponse, ReplaceAlertRuleStatusApiArg>({
-        query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}/status`,
-          method: 'PUT',
-          body: queryArg.alertRule,
-          params: {
-            pretty: queryArg.pretty,
-            dryRun: queryArg.dryRun,
-            fieldManager: queryArg.fieldManager,
-            fieldValidation: queryArg.fieldValidation,
-          },
-        }),
-        invalidatesTags: ['AlertRule'],
-      }),
-      updateAlertRuleStatus: build.mutation<UpdateAlertRuleStatusApiResponse, UpdateAlertRuleStatusApiArg>({
-        query: (queryArg) => ({
-          url: `/alertrules/${queryArg.name}/status`,
-          method: 'PATCH',
-          body: queryArg.patch,
-          params: {
-            pretty: queryArg.pretty,
-            dryRun: queryArg.dryRun,
-            fieldManager: queryArg.fieldManager,
-            fieldValidation: queryArg.fieldValidation,
-            force: queryArg.force,
-          },
-        }),
-        invalidatesTags: ['AlertRule'],
-      }),
-      listRecordingRule: build.query<ListRecordingRuleApiResponse, ListRecordingRuleApiArg>({
-        query: (queryArg) => ({
-          url: `/recordingrules`,
+          url: `/routingtrees`,
           params: {
             pretty: queryArg.pretty,
             allowWatchBookmarks: queryArg.allowWatchBookmarks,
@@ -174,13 +136,13 @@ const injectedRtkApi = api
             watch: queryArg.watch,
           },
         }),
-        providesTags: ['RecordingRule'],
+        providesTags: ['RoutingTree'],
       }),
-      createRecordingRule: build.mutation<CreateRecordingRuleApiResponse, CreateRecordingRuleApiArg>({
+      createRoutingTree: build.mutation<CreateRoutingTreeApiResponse, CreateRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules`,
+          url: `/routingtrees`,
           method: 'POST',
-          body: queryArg.recordingRule,
+          body: queryArg.routingTree,
           params: {
             pretty: queryArg.pretty,
             dryRun: queryArg.dryRun,
@@ -188,14 +150,14 @@ const injectedRtkApi = api
             fieldValidation: queryArg.fieldValidation,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['RoutingTree'],
       }),
-      deletecollectionRecordingRule: build.mutation<
-        DeletecollectionRecordingRuleApiResponse,
-        DeletecollectionRecordingRuleApiArg
+      deletecollectionRoutingTree: build.mutation<
+        DeletecollectionRoutingTreeApiResponse,
+        DeletecollectionRoutingTreeApiArg
       >({
         query: (queryArg) => ({
-          url: `/recordingrules`,
+          url: `/routingtrees`,
           method: 'DELETE',
           params: {
             pretty: queryArg.pretty,
@@ -214,22 +176,22 @@ const injectedRtkApi = api
             timeoutSeconds: queryArg.timeoutSeconds,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['RoutingTree'],
       }),
-      getRecordingRule: build.query<GetRecordingRuleApiResponse, GetRecordingRuleApiArg>({
+      getRoutingTree: build.query<GetRoutingTreeApiResponse, GetRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}`,
+          url: `/routingtrees/${queryArg.name}`,
           params: {
             pretty: queryArg.pretty,
           },
         }),
-        providesTags: ['RecordingRule'],
+        providesTags: ['RoutingTree'],
       }),
-      replaceRecordingRule: build.mutation<ReplaceRecordingRuleApiResponse, ReplaceRecordingRuleApiArg>({
+      replaceRoutingTree: build.mutation<ReplaceRoutingTreeApiResponse, ReplaceRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}`,
+          url: `/routingtrees/${queryArg.name}`,
           method: 'PUT',
-          body: queryArg.recordingRule,
+          body: queryArg.routingTree,
           params: {
             pretty: queryArg.pretty,
             dryRun: queryArg.dryRun,
@@ -237,11 +199,11 @@ const injectedRtkApi = api
             fieldValidation: queryArg.fieldValidation,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['RoutingTree'],
       }),
-      deleteRecordingRule: build.mutation<DeleteRecordingRuleApiResponse, DeleteRecordingRuleApiArg>({
+      deleteRoutingTree: build.mutation<DeleteRoutingTreeApiResponse, DeleteRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}`,
+          url: `/routingtrees/${queryArg.name}`,
           method: 'DELETE',
           params: {
             pretty: queryArg.pretty,
@@ -252,11 +214,11 @@ const injectedRtkApi = api
             propagationPolicy: queryArg.propagationPolicy,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['RoutingTree'],
       }),
-      updateRecordingRule: build.mutation<UpdateRecordingRuleApiResponse, UpdateRecordingRuleApiArg>({
+      updateRoutingTree: build.mutation<UpdateRoutingTreeApiResponse, UpdateRoutingTreeApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}`,
+          url: `/routingtrees/${queryArg.name}`,
           method: 'PATCH',
           body: queryArg.patch,
           params: {
@@ -267,25 +229,81 @@ const injectedRtkApi = api
             force: queryArg.force,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['RoutingTree'],
       }),
-      getRecordingRuleStatus: build.query<GetRecordingRuleStatusApiResponse, GetRecordingRuleStatusApiArg>({
+      listTemplateGroup: build.query<ListTemplateGroupApiResponse, ListTemplateGroupApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}/status`,
+          url: `/templategroups`,
           params: {
             pretty: queryArg.pretty,
+            allowWatchBookmarks: queryArg.allowWatchBookmarks,
+            continue: queryArg['continue'],
+            fieldSelector: queryArg.fieldSelector,
+            labelSelector: queryArg.labelSelector,
+            limit: queryArg.limit,
+            resourceVersion: queryArg.resourceVersion,
+            resourceVersionMatch: queryArg.resourceVersionMatch,
+            sendInitialEvents: queryArg.sendInitialEvents,
+            timeoutSeconds: queryArg.timeoutSeconds,
+            watch: queryArg.watch,
           },
         }),
-        providesTags: ['RecordingRule'],
+        providesTags: ['TemplateGroup'],
       }),
-      replaceRecordingRuleStatus: build.mutation<
-        ReplaceRecordingRuleStatusApiResponse,
-        ReplaceRecordingRuleStatusApiArg
+      createTemplateGroup: build.mutation<CreateTemplateGroupApiResponse, CreateTemplateGroupApiArg>({
+        query: (queryArg) => ({
+          url: `/templategroups`,
+          method: 'POST',
+          body: queryArg.templateGroup,
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            fieldManager: queryArg.fieldManager,
+            fieldValidation: queryArg.fieldValidation,
+          },
+        }),
+        invalidatesTags: ['TemplateGroup'],
+      }),
+      deletecollectionTemplateGroup: build.mutation<
+        DeletecollectionTemplateGroupApiResponse,
+        DeletecollectionTemplateGroupApiArg
       >({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}/status`,
+          url: `/templategroups`,
+          method: 'DELETE',
+          params: {
+            pretty: queryArg.pretty,
+            continue: queryArg['continue'],
+            dryRun: queryArg.dryRun,
+            fieldSelector: queryArg.fieldSelector,
+            gracePeriodSeconds: queryArg.gracePeriodSeconds,
+            ignoreStoreReadErrorWithClusterBreakingPotential: queryArg.ignoreStoreReadErrorWithClusterBreakingPotential,
+            labelSelector: queryArg.labelSelector,
+            limit: queryArg.limit,
+            orphanDependents: queryArg.orphanDependents,
+            propagationPolicy: queryArg.propagationPolicy,
+            resourceVersion: queryArg.resourceVersion,
+            resourceVersionMatch: queryArg.resourceVersionMatch,
+            sendInitialEvents: queryArg.sendInitialEvents,
+            timeoutSeconds: queryArg.timeoutSeconds,
+          },
+        }),
+        invalidatesTags: ['TemplateGroup'],
+      }),
+      getTemplateGroup: build.query<GetTemplateGroupApiResponse, GetTemplateGroupApiArg>({
+        query: (queryArg) => ({
+          url: `/templategroups/${queryArg.name}`,
+          params: {
+            pretty: queryArg.pretty,
+          },
+        }),
+        providesTags: ['TemplateGroup'],
+      }),
+      replaceTemplateGroup: build.mutation<ReplaceTemplateGroupApiResponse, ReplaceTemplateGroupApiArg>({
+        query: (queryArg) => ({
+          url: `/templategroups/${queryArg.name}`,
           method: 'PUT',
-          body: queryArg.recordingRule,
+          body: queryArg.templateGroup,
           params: {
             pretty: queryArg.pretty,
             dryRun: queryArg.dryRun,
@@ -293,11 +311,26 @@ const injectedRtkApi = api
             fieldValidation: queryArg.fieldValidation,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['TemplateGroup'],
       }),
-      updateRecordingRuleStatus: build.mutation<UpdateRecordingRuleStatusApiResponse, UpdateRecordingRuleStatusApiArg>({
+      deleteTemplateGroup: build.mutation<DeleteTemplateGroupApiResponse, DeleteTemplateGroupApiArg>({
         query: (queryArg) => ({
-          url: `/recordingrules/${queryArg.name}/status`,
+          url: `/templategroups/${queryArg.name}`,
+          method: 'DELETE',
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            gracePeriodSeconds: queryArg.gracePeriodSeconds,
+            ignoreStoreReadErrorWithClusterBreakingPotential: queryArg.ignoreStoreReadErrorWithClusterBreakingPotential,
+            orphanDependents: queryArg.orphanDependents,
+            propagationPolicy: queryArg.propagationPolicy,
+          },
+        }),
+        invalidatesTags: ['TemplateGroup'],
+      }),
+      updateTemplateGroup: build.mutation<UpdateTemplateGroupApiResponse, UpdateTemplateGroupApiArg>({
+        query: (queryArg) => ({
+          url: `/templategroups/${queryArg.name}`,
           method: 'PATCH',
           body: queryArg.patch,
           params: {
@@ -308,16 +341,128 @@ const injectedRtkApi = api
             force: queryArg.force,
           },
         }),
-        invalidatesTags: ['RecordingRule'],
+        invalidatesTags: ['TemplateGroup'],
+      }),
+      listTimeInterval: build.query<ListTimeIntervalApiResponse, ListTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals`,
+          params: {
+            pretty: queryArg.pretty,
+            allowWatchBookmarks: queryArg.allowWatchBookmarks,
+            continue: queryArg['continue'],
+            fieldSelector: queryArg.fieldSelector,
+            labelSelector: queryArg.labelSelector,
+            limit: queryArg.limit,
+            resourceVersion: queryArg.resourceVersion,
+            resourceVersionMatch: queryArg.resourceVersionMatch,
+            sendInitialEvents: queryArg.sendInitialEvents,
+            timeoutSeconds: queryArg.timeoutSeconds,
+            watch: queryArg.watch,
+          },
+        }),
+        providesTags: ['TimeInterval'],
+      }),
+      createTimeInterval: build.mutation<CreateTimeIntervalApiResponse, CreateTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals`,
+          method: 'POST',
+          body: queryArg.timeInterval,
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            fieldManager: queryArg.fieldManager,
+            fieldValidation: queryArg.fieldValidation,
+          },
+        }),
+        invalidatesTags: ['TimeInterval'],
+      }),
+      deletecollectionTimeInterval: build.mutation<
+        DeletecollectionTimeIntervalApiResponse,
+        DeletecollectionTimeIntervalApiArg
+      >({
+        query: (queryArg) => ({
+          url: `/timeintervals`,
+          method: 'DELETE',
+          params: {
+            pretty: queryArg.pretty,
+            continue: queryArg['continue'],
+            dryRun: queryArg.dryRun,
+            fieldSelector: queryArg.fieldSelector,
+            gracePeriodSeconds: queryArg.gracePeriodSeconds,
+            ignoreStoreReadErrorWithClusterBreakingPotential: queryArg.ignoreStoreReadErrorWithClusterBreakingPotential,
+            labelSelector: queryArg.labelSelector,
+            limit: queryArg.limit,
+            orphanDependents: queryArg.orphanDependents,
+            propagationPolicy: queryArg.propagationPolicy,
+            resourceVersion: queryArg.resourceVersion,
+            resourceVersionMatch: queryArg.resourceVersionMatch,
+            sendInitialEvents: queryArg.sendInitialEvents,
+            timeoutSeconds: queryArg.timeoutSeconds,
+          },
+        }),
+        invalidatesTags: ['TimeInterval'],
+      }),
+      getTimeInterval: build.query<GetTimeIntervalApiResponse, GetTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals/${queryArg.name}`,
+          params: {
+            pretty: queryArg.pretty,
+          },
+        }),
+        providesTags: ['TimeInterval'],
+      }),
+      replaceTimeInterval: build.mutation<ReplaceTimeIntervalApiResponse, ReplaceTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals/${queryArg.name}`,
+          method: 'PUT',
+          body: queryArg.timeInterval,
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            fieldManager: queryArg.fieldManager,
+            fieldValidation: queryArg.fieldValidation,
+          },
+        }),
+        invalidatesTags: ['TimeInterval'],
+      }),
+      deleteTimeInterval: build.mutation<DeleteTimeIntervalApiResponse, DeleteTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals/${queryArg.name}`,
+          method: 'DELETE',
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            gracePeriodSeconds: queryArg.gracePeriodSeconds,
+            ignoreStoreReadErrorWithClusterBreakingPotential: queryArg.ignoreStoreReadErrorWithClusterBreakingPotential,
+            orphanDependents: queryArg.orphanDependents,
+            propagationPolicy: queryArg.propagationPolicy,
+          },
+        }),
+        invalidatesTags: ['TimeInterval'],
+      }),
+      updateTimeInterval: build.mutation<UpdateTimeIntervalApiResponse, UpdateTimeIntervalApiArg>({
+        query: (queryArg) => ({
+          url: `/timeintervals/${queryArg.name}`,
+          method: 'PATCH',
+          body: queryArg.patch,
+          params: {
+            pretty: queryArg.pretty,
+            dryRun: queryArg.dryRun,
+            fieldManager: queryArg.fieldManager,
+            fieldValidation: queryArg.fieldValidation,
+            force: queryArg.force,
+          },
+        }),
+        invalidatesTags: ['TimeInterval'],
       }),
     }),
     overrideExisting: false,
   });
-export { injectedRtkApi as rulesAPI };
+export { injectedRtkApi as generatedAPI };
 export type GetApiResourcesApiResponse = /** status 200 OK */ ApiResourceList;
 export type GetApiResourcesApiArg = void;
-export type ListAlertRuleApiResponse = /** status 200 OK */ AlertRuleList;
-export type ListAlertRuleApiArg = {
+export type ListReceiverApiResponse = /** status 200 OK */ ReceiverList;
+export type ListReceiverApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** allowWatchBookmarks requests watch events with type "BOOKMARK". Servers that do not implement bookmarks may ignore this flag and bookmarks are sent at the server's discretion. Clients should not assume bookmarks are returned at any specific interval, nor may they assume the server will send any BOOKMARK event during a session. If this is not a watch, this field is ignored. */
@@ -361,11 +506,11 @@ export type ListAlertRuleApiArg = {
   /** Watch for changes to the described resources and return them as a stream of add, update, and remove notifications. Specify resourceVersion. */
   watch?: boolean;
 };
-export type CreateAlertRuleApiResponse = /** status 200 OK */
-  | AlertRule
-  | /** status 201 Created */ AlertRule
-  | /** status 202 Accepted */ AlertRule;
-export type CreateAlertRuleApiArg = {
+export type CreateReceiverApiResponse = /** status 200 OK */
+  | Receiver
+  | /** status 201 Created */ Receiver
+  | /** status 202 Accepted */ Receiver;
+export type CreateReceiverApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
@@ -374,10 +519,10 @@ export type CreateAlertRuleApiArg = {
   fieldManager?: string;
   /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
   fieldValidation?: string;
-  alertRule: AlertRule;
+  receiver: Receiver;
 };
-export type DeletecollectionAlertRuleApiResponse = /** status 200 OK */ Status;
-export type DeletecollectionAlertRuleApiArg = {
+export type DeletecollectionReceiverApiResponse = /** status 200 OK */ Status;
+export type DeletecollectionReceiverApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
@@ -427,16 +572,16 @@ export type DeletecollectionAlertRuleApiArg = {
   /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
   timeoutSeconds?: number;
 };
-export type GetAlertRuleApiResponse = /** status 200 OK */ AlertRule;
-export type GetAlertRuleApiArg = {
-  /** name of the AlertRule */
+export type GetReceiverApiResponse = /** status 200 OK */ Receiver;
+export type GetReceiverApiArg = {
+  /** name of the Receiver */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
 };
-export type ReplaceAlertRuleApiResponse = /** status 200 OK */ AlertRule | /** status 201 Created */ AlertRule;
-export type ReplaceAlertRuleApiArg = {
-  /** name of the AlertRule */
+export type ReplaceReceiverApiResponse = /** status 200 OK */ Receiver | /** status 201 Created */ Receiver;
+export type ReplaceReceiverApiArg = {
+  /** name of the Receiver */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -446,11 +591,11 @@ export type ReplaceAlertRuleApiArg = {
   fieldManager?: string;
   /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
   fieldValidation?: string;
-  alertRule: AlertRule;
+  receiver: Receiver;
 };
-export type DeleteAlertRuleApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
-export type DeleteAlertRuleApiArg = {
-  /** name of the AlertRule */
+export type DeleteReceiverApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
+export type DeleteReceiverApiArg = {
+  /** name of the Receiver */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -465,9 +610,9 @@ export type DeleteAlertRuleApiArg = {
   /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
   propagationPolicy?: string;
 };
-export type UpdateAlertRuleApiResponse = /** status 200 OK */ AlertRule | /** status 201 Created */ AlertRule;
-export type UpdateAlertRuleApiArg = {
-  /** name of the AlertRule */
+export type UpdateReceiverApiResponse = /** status 200 OK */ Receiver | /** status 201 Created */ Receiver;
+export type UpdateReceiverApiArg = {
+  /** name of the Receiver */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -481,45 +626,8 @@ export type UpdateAlertRuleApiArg = {
   force?: boolean;
   patch: Patch;
 };
-export type GetAlertRuleStatusApiResponse = /** status 200 OK */ AlertRule;
-export type GetAlertRuleStatusApiArg = {
-  /** name of the AlertRule */
-  name: string;
-  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
-  pretty?: string;
-};
-export type ReplaceAlertRuleStatusApiResponse = /** status 200 OK */ AlertRule | /** status 201 Created */ AlertRule;
-export type ReplaceAlertRuleStatusApiArg = {
-  /** name of the AlertRule */
-  name: string;
-  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
-  pretty?: string;
-  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
-  dryRun?: string;
-  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. */
-  fieldManager?: string;
-  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
-  fieldValidation?: string;
-  alertRule: AlertRule;
-};
-export type UpdateAlertRuleStatusApiResponse = /** status 200 OK */ AlertRule | /** status 201 Created */ AlertRule;
-export type UpdateAlertRuleStatusApiArg = {
-  /** name of the AlertRule */
-  name: string;
-  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
-  pretty?: string;
-  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
-  dryRun?: string;
-  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. This field is required for apply requests (application/apply-patch) but optional for non-apply patch types (JsonPatch, MergePatch, StrategicMergePatch). */
-  fieldManager?: string;
-  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
-  fieldValidation?: string;
-  /** Force is going to "force" Apply requests. It means user will re-acquire conflicting fields owned by other people. Force flag must be unset for non-apply patch requests. */
-  force?: boolean;
-  patch: Patch;
-};
-export type ListRecordingRuleApiResponse = /** status 200 OK */ RecordingRuleList;
-export type ListRecordingRuleApiArg = {
+export type ListRoutingTreeApiResponse = /** status 200 OK */ RoutingTreeList;
+export type ListRoutingTreeApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** allowWatchBookmarks requests watch events with type "BOOKMARK". Servers that do not implement bookmarks may ignore this flag and bookmarks are sent at the server's discretion. Clients should not assume bookmarks are returned at any specific interval, nor may they assume the server will send any BOOKMARK event during a session. If this is not a watch, this field is ignored. */
@@ -563,11 +671,11 @@ export type ListRecordingRuleApiArg = {
   /** Watch for changes to the described resources and return them as a stream of add, update, and remove notifications. Specify resourceVersion. */
   watch?: boolean;
 };
-export type CreateRecordingRuleApiResponse = /** status 200 OK */
-  | RecordingRule
-  | /** status 201 Created */ RecordingRule
-  | /** status 202 Accepted */ RecordingRule;
-export type CreateRecordingRuleApiArg = {
+export type CreateRoutingTreeApiResponse = /** status 200 OK */
+  | RoutingTree
+  | /** status 201 Created */ RoutingTree
+  | /** status 202 Accepted */ RoutingTree;
+export type CreateRoutingTreeApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
@@ -576,10 +684,10 @@ export type CreateRecordingRuleApiArg = {
   fieldManager?: string;
   /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
   fieldValidation?: string;
-  recordingRule: RecordingRule;
+  routingTree: RoutingTree;
 };
-export type DeletecollectionRecordingRuleApiResponse = /** status 200 OK */ Status;
-export type DeletecollectionRecordingRuleApiArg = {
+export type DeletecollectionRoutingTreeApiResponse = /** status 200 OK */ Status;
+export type DeletecollectionRoutingTreeApiArg = {
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
   /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
@@ -629,18 +737,16 @@ export type DeletecollectionRecordingRuleApiArg = {
   /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
   timeoutSeconds?: number;
 };
-export type GetRecordingRuleApiResponse = /** status 200 OK */ RecordingRule;
-export type GetRecordingRuleApiArg = {
-  /** name of the RecordingRule */
+export type GetRoutingTreeApiResponse = /** status 200 OK */ RoutingTree;
+export type GetRoutingTreeApiArg = {
+  /** name of the RoutingTree */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
 };
-export type ReplaceRecordingRuleApiResponse = /** status 200 OK */
-  | RecordingRule
-  | /** status 201 Created */ RecordingRule;
-export type ReplaceRecordingRuleApiArg = {
-  /** name of the RecordingRule */
+export type ReplaceRoutingTreeApiResponse = /** status 200 OK */ RoutingTree | /** status 201 Created */ RoutingTree;
+export type ReplaceRoutingTreeApiArg = {
+  /** name of the RoutingTree */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -650,11 +756,11 @@ export type ReplaceRecordingRuleApiArg = {
   fieldManager?: string;
   /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
   fieldValidation?: string;
-  recordingRule: RecordingRule;
+  routingTree: RoutingTree;
 };
-export type DeleteRecordingRuleApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
-export type DeleteRecordingRuleApiArg = {
-  /** name of the RecordingRule */
+export type DeleteRoutingTreeApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
+export type DeleteRoutingTreeApiArg = {
+  /** name of the RoutingTree */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -669,11 +775,9 @@ export type DeleteRecordingRuleApiArg = {
   /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
   propagationPolicy?: string;
 };
-export type UpdateRecordingRuleApiResponse = /** status 200 OK */
-  | RecordingRule
-  | /** status 201 Created */ RecordingRule;
-export type UpdateRecordingRuleApiArg = {
-  /** name of the RecordingRule */
+export type UpdateRoutingTreeApiResponse = /** status 200 OK */ RoutingTree | /** status 201 Created */ RoutingTree;
+export type UpdateRoutingTreeApiArg = {
+  /** name of the RoutingTree */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -687,18 +791,129 @@ export type UpdateRecordingRuleApiArg = {
   force?: boolean;
   patch: Patch;
 };
-export type GetRecordingRuleStatusApiResponse = /** status 200 OK */ RecordingRule;
-export type GetRecordingRuleStatusApiArg = {
-  /** name of the RecordingRule */
+export type ListTemplateGroupApiResponse = /** status 200 OK */ TemplateGroupList;
+export type ListTemplateGroupApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** allowWatchBookmarks requests watch events with type "BOOKMARK". Servers that do not implement bookmarks may ignore this flag and bookmarks are sent at the server's discretion. Clients should not assume bookmarks are returned at any specific interval, nor may they assume the server will send any BOOKMARK event during a session. If this is not a watch, this field is ignored. */
+  allowWatchBookmarks?: boolean;
+  /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
+    
+    This field is not supported when watch is true. Clients may start a watch from the last resourceVersion value returned by the server and not miss any modifications. */
+  continue?: string;
+  /** A selector to restrict the list of returned objects by their fields. Defaults to everything. */
+  fieldSelector?: string;
+  /** A selector to restrict the list of returned objects by their labels. Defaults to everything. */
+  labelSelector?: string;
+  /** limit is a maximum number of responses to return for a list call. If more items exist, the server will set the `continue` field on the list metadata to a value that can be used with the same initial query to retrieve the next set of results. Setting a limit may return fewer than the requested amount of items (up to zero items) in the event all requested objects are filtered out and clients should only use the presence of the continue field to determine whether more results are available. Servers may choose not to support the limit argument and will return all of the available results. If limit is specified and the continue field is empty, clients may assume that no more results are available. This field is not supported if watch is true.
+    
+    The server guarantees that the objects returned when using continue will be identical to issuing a single list call without a limit - that is, no objects created, modified, or deleted after the first request is issued will be included in any subsequent continued requests. This is sometimes referred to as a consistent snapshot, and ensures that a client that is using limit to receive smaller chunks of a very large result can ensure they see all possible objects. If objects are updated during a chunked list the version of the object that was present at the time the first list result was calculated is returned. */
+  limit?: number;
+  /** resourceVersion sets a constraint on what resource versions a request may be served from. See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersion?: string;
+  /** resourceVersionMatch determines how resourceVersion is applied to list calls. It is highly recommended that resourceVersionMatch be set for list calls where resourceVersion is set See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersionMatch?: string;
+  /** `sendInitialEvents=true` may be set together with `watch=true`. In that case, the watch stream will begin with synthetic events to produce the current state of objects in the collection. Once all such events have been sent, a synthetic "Bookmark" event  will be sent. The bookmark will report the ResourceVersion (RV) corresponding to the set of objects, and be marked with `"k8s.io/initial-events-end": "true"` annotation. Afterwards, the watch stream will proceed as usual, sending watch events corresponding to changes (subsequent to the RV) to objects watched.
+    
+    When `sendInitialEvents` option is set, we require `resourceVersionMatch` option to also be set. The semantic of the watch request is as following: - `resourceVersionMatch` = NotOlderThan
+      is interpreted as "data at least as new as the provided `resourceVersion`"
+      and the bookmark event is send when the state is synced
+      to a `resourceVersion` at least as fresh as the one provided by the ListOptions.
+      If `resourceVersion` is unset, this is interpreted as "consistent read" and the
+      bookmark event is send when the state is synced at least to the moment
+      when request started being processed.
+    - `resourceVersionMatch` set to any other value or unset
+      Invalid error is returned.
+    
+    Defaults to true if `resourceVersion=""` or `resourceVersion="0"` (for backward compatibility reasons) and to false otherwise. */
+  sendInitialEvents?: boolean;
+  /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
+  timeoutSeconds?: number;
+  /** Watch for changes to the described resources and return them as a stream of add, update, and remove notifications. Specify resourceVersion. */
+  watch?: boolean;
+};
+export type CreateTemplateGroupApiResponse = /** status 200 OK */
+  | TemplateGroup
+  | /** status 201 Created */ TemplateGroup
+  | /** status 202 Accepted */ TemplateGroup;
+export type CreateTemplateGroupApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. */
+  fieldManager?: string;
+  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
+  fieldValidation?: string;
+  templateGroup: TemplateGroup;
+};
+export type DeletecollectionTemplateGroupApiResponse = /** status 200 OK */ Status;
+export type DeletecollectionTemplateGroupApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
+    
+    This field is not supported when watch is true. Clients may start a watch from the last resourceVersion value returned by the server and not miss any modifications. */
+  continue?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** A selector to restrict the list of returned objects by their fields. Defaults to everything. */
+  fieldSelector?: string;
+  /** The duration in seconds before the object should be deleted. Value must be non-negative integer. The value zero indicates delete immediately. If this value is nil, the default grace period for the specified type will be used. Defaults to a per object value if not specified. zero means delete immediately. */
+  gracePeriodSeconds?: number;
+  /** if set to true, it will trigger an unsafe deletion of the resource in case the normal deletion flow fails with a corrupt object error. A resource is considered corrupt if it can not be retrieved from the underlying storage successfully because of a) its data can not be transformed e.g. decryption failure, or b) it fails to decode into an object. NOTE: unsafe deletion ignores finalizer constraints, skips precondition checks, and removes the object from the storage. WARNING: This may potentially break the cluster if the workload associated with the resource being unsafe-deleted relies on normal deletion flow. Use only if you REALLY know what you are doing. The default value is false, and the user must opt in to enable it */
+  ignoreStoreReadErrorWithClusterBreakingPotential?: boolean;
+  /** A selector to restrict the list of returned objects by their labels. Defaults to everything. */
+  labelSelector?: string;
+  /** limit is a maximum number of responses to return for a list call. If more items exist, the server will set the `continue` field on the list metadata to a value that can be used with the same initial query to retrieve the next set of results. Setting a limit may return fewer than the requested amount of items (up to zero items) in the event all requested objects are filtered out and clients should only use the presence of the continue field to determine whether more results are available. Servers may choose not to support the limit argument and will return all of the available results. If limit is specified and the continue field is empty, clients may assume that no more results are available. This field is not supported if watch is true.
+    
+    The server guarantees that the objects returned when using continue will be identical to issuing a single list call without a limit - that is, no objects created, modified, or deleted after the first request is issued will be included in any subsequent continued requests. This is sometimes referred to as a consistent snapshot, and ensures that a client that is using limit to receive smaller chunks of a very large result can ensure they see all possible objects. If objects are updated during a chunked list the version of the object that was present at the time the first list result was calculated is returned. */
+  limit?: number;
+  /** Deprecated: please use the PropagationPolicy, this field will be deprecated in 1.7. Should the dependent objects be orphaned. If true/false, the "orphan" finalizer will be added to/removed from the object's finalizers list. Either this field or PropagationPolicy may be set, but not both. */
+  orphanDependents?: boolean;
+  /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
+  propagationPolicy?: string;
+  /** resourceVersion sets a constraint on what resource versions a request may be served from. See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersion?: string;
+  /** resourceVersionMatch determines how resourceVersion is applied to list calls. It is highly recommended that resourceVersionMatch be set for list calls where resourceVersion is set See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersionMatch?: string;
+  /** `sendInitialEvents=true` may be set together with `watch=true`. In that case, the watch stream will begin with synthetic events to produce the current state of objects in the collection. Once all such events have been sent, a synthetic "Bookmark" event  will be sent. The bookmark will report the ResourceVersion (RV) corresponding to the set of objects, and be marked with `"k8s.io/initial-events-end": "true"` annotation. Afterwards, the watch stream will proceed as usual, sending watch events corresponding to changes (subsequent to the RV) to objects watched.
+    
+    When `sendInitialEvents` option is set, we require `resourceVersionMatch` option to also be set. The semantic of the watch request is as following: - `resourceVersionMatch` = NotOlderThan
+      is interpreted as "data at least as new as the provided `resourceVersion`"
+      and the bookmark event is send when the state is synced
+      to a `resourceVersion` at least as fresh as the one provided by the ListOptions.
+      If `resourceVersion` is unset, this is interpreted as "consistent read" and the
+      bookmark event is send when the state is synced at least to the moment
+      when request started being processed.
+    - `resourceVersionMatch` set to any other value or unset
+      Invalid error is returned.
+    
+    Defaults to true if `resourceVersion=""` or `resourceVersion="0"` (for backward compatibility reasons) and to false otherwise. */
+  sendInitialEvents?: boolean;
+  /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
+  timeoutSeconds?: number;
+};
+export type GetTemplateGroupApiResponse = /** status 200 OK */ TemplateGroup;
+export type GetTemplateGroupApiArg = {
+  /** name of the TemplateGroup */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
 };
-export type ReplaceRecordingRuleStatusApiResponse = /** status 200 OK */
-  | RecordingRule
-  | /** status 201 Created */ RecordingRule;
-export type ReplaceRecordingRuleStatusApiArg = {
-  /** name of the RecordingRule */
+export type ReplaceTemplateGroupApiResponse = /** status 200 OK */
+  | TemplateGroup
+  | /** status 201 Created */ TemplateGroup;
+export type ReplaceTemplateGroupApiArg = {
+  /** name of the TemplateGroup */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -708,13 +923,195 @@ export type ReplaceRecordingRuleStatusApiArg = {
   fieldManager?: string;
   /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
   fieldValidation?: string;
-  recordingRule: RecordingRule;
+  templateGroup: TemplateGroup;
 };
-export type UpdateRecordingRuleStatusApiResponse = /** status 200 OK */
-  | RecordingRule
-  | /** status 201 Created */ RecordingRule;
-export type UpdateRecordingRuleStatusApiArg = {
-  /** name of the RecordingRule */
+export type DeleteTemplateGroupApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
+export type DeleteTemplateGroupApiArg = {
+  /** name of the TemplateGroup */
+  name: string;
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** The duration in seconds before the object should be deleted. Value must be non-negative integer. The value zero indicates delete immediately. If this value is nil, the default grace period for the specified type will be used. Defaults to a per object value if not specified. zero means delete immediately. */
+  gracePeriodSeconds?: number;
+  /** if set to true, it will trigger an unsafe deletion of the resource in case the normal deletion flow fails with a corrupt object error. A resource is considered corrupt if it can not be retrieved from the underlying storage successfully because of a) its data can not be transformed e.g. decryption failure, or b) it fails to decode into an object. NOTE: unsafe deletion ignores finalizer constraints, skips precondition checks, and removes the object from the storage. WARNING: This may potentially break the cluster if the workload associated with the resource being unsafe-deleted relies on normal deletion flow. Use only if you REALLY know what you are doing. The default value is false, and the user must opt in to enable it */
+  ignoreStoreReadErrorWithClusterBreakingPotential?: boolean;
+  /** Deprecated: please use the PropagationPolicy, this field will be deprecated in 1.7. Should the dependent objects be orphaned. If true/false, the "orphan" finalizer will be added to/removed from the object's finalizers list. Either this field or PropagationPolicy may be set, but not both. */
+  orphanDependents?: boolean;
+  /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
+  propagationPolicy?: string;
+};
+export type UpdateTemplateGroupApiResponse = /** status 200 OK */
+  | TemplateGroup
+  | /** status 201 Created */ TemplateGroup;
+export type UpdateTemplateGroupApiArg = {
+  /** name of the TemplateGroup */
+  name: string;
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. This field is required for apply requests (application/apply-patch) but optional for non-apply patch types (JsonPatch, MergePatch, StrategicMergePatch). */
+  fieldManager?: string;
+  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
+  fieldValidation?: string;
+  /** Force is going to "force" Apply requests. It means user will re-acquire conflicting fields owned by other people. Force flag must be unset for non-apply patch requests. */
+  force?: boolean;
+  patch: Patch;
+};
+export type ListTimeIntervalApiResponse = /** status 200 OK */ TimeIntervalList;
+export type ListTimeIntervalApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** allowWatchBookmarks requests watch events with type "BOOKMARK". Servers that do not implement bookmarks may ignore this flag and bookmarks are sent at the server's discretion. Clients should not assume bookmarks are returned at any specific interval, nor may they assume the server will send any BOOKMARK event during a session. If this is not a watch, this field is ignored. */
+  allowWatchBookmarks?: boolean;
+  /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
+    
+    This field is not supported when watch is true. Clients may start a watch from the last resourceVersion value returned by the server and not miss any modifications. */
+  continue?: string;
+  /** A selector to restrict the list of returned objects by their fields. Defaults to everything. */
+  fieldSelector?: string;
+  /** A selector to restrict the list of returned objects by their labels. Defaults to everything. */
+  labelSelector?: string;
+  /** limit is a maximum number of responses to return for a list call. If more items exist, the server will set the `continue` field on the list metadata to a value that can be used with the same initial query to retrieve the next set of results. Setting a limit may return fewer than the requested amount of items (up to zero items) in the event all requested objects are filtered out and clients should only use the presence of the continue field to determine whether more results are available. Servers may choose not to support the limit argument and will return all of the available results. If limit is specified and the continue field is empty, clients may assume that no more results are available. This field is not supported if watch is true.
+    
+    The server guarantees that the objects returned when using continue will be identical to issuing a single list call without a limit - that is, no objects created, modified, or deleted after the first request is issued will be included in any subsequent continued requests. This is sometimes referred to as a consistent snapshot, and ensures that a client that is using limit to receive smaller chunks of a very large result can ensure they see all possible objects. If objects are updated during a chunked list the version of the object that was present at the time the first list result was calculated is returned. */
+  limit?: number;
+  /** resourceVersion sets a constraint on what resource versions a request may be served from. See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersion?: string;
+  /** resourceVersionMatch determines how resourceVersion is applied to list calls. It is highly recommended that resourceVersionMatch be set for list calls where resourceVersion is set See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersionMatch?: string;
+  /** `sendInitialEvents=true` may be set together with `watch=true`. In that case, the watch stream will begin with synthetic events to produce the current state of objects in the collection. Once all such events have been sent, a synthetic "Bookmark" event  will be sent. The bookmark will report the ResourceVersion (RV) corresponding to the set of objects, and be marked with `"k8s.io/initial-events-end": "true"` annotation. Afterwards, the watch stream will proceed as usual, sending watch events corresponding to changes (subsequent to the RV) to objects watched.
+    
+    When `sendInitialEvents` option is set, we require `resourceVersionMatch` option to also be set. The semantic of the watch request is as following: - `resourceVersionMatch` = NotOlderThan
+      is interpreted as "data at least as new as the provided `resourceVersion`"
+      and the bookmark event is send when the state is synced
+      to a `resourceVersion` at least as fresh as the one provided by the ListOptions.
+      If `resourceVersion` is unset, this is interpreted as "consistent read" and the
+      bookmark event is send when the state is synced at least to the moment
+      when request started being processed.
+    - `resourceVersionMatch` set to any other value or unset
+      Invalid error is returned.
+    
+    Defaults to true if `resourceVersion=""` or `resourceVersion="0"` (for backward compatibility reasons) and to false otherwise. */
+  sendInitialEvents?: boolean;
+  /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
+  timeoutSeconds?: number;
+  /** Watch for changes to the described resources and return them as a stream of add, update, and remove notifications. Specify resourceVersion. */
+  watch?: boolean;
+};
+export type CreateTimeIntervalApiResponse = /** status 200 OK */
+  | TimeInterval
+  | /** status 201 Created */ TimeInterval
+  | /** status 202 Accepted */ TimeInterval;
+export type CreateTimeIntervalApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. */
+  fieldManager?: string;
+  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
+  fieldValidation?: string;
+  timeInterval: TimeInterval;
+};
+export type DeletecollectionTimeIntervalApiResponse = /** status 200 OK */ Status;
+export type DeletecollectionTimeIntervalApiArg = {
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** The continue option should be set when retrieving more results from the server. Since this value is server defined, clients may only use the continue value from a previous query result with identical query parameters (except for the value of continue) and the server may reject a continue value it does not recognize. If the specified continue value is no longer valid whether due to expiration (generally five to fifteen minutes) or a configuration change on the server, the server will respond with a 410 ResourceExpired error together with a continue token. If the client needs a consistent list, it must restart their list without the continue field. Otherwise, the client may send another list request with the token received with the 410 error, the server will respond with a list starting from the next key, but from the latest snapshot, which is inconsistent from the previous list results - objects that are created, modified, or deleted after the first list request will be included in the response, as long as their keys are after the "next key".
+    
+    This field is not supported when watch is true. Clients may start a watch from the last resourceVersion value returned by the server and not miss any modifications. */
+  continue?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** A selector to restrict the list of returned objects by their fields. Defaults to everything. */
+  fieldSelector?: string;
+  /** The duration in seconds before the object should be deleted. Value must be non-negative integer. The value zero indicates delete immediately. If this value is nil, the default grace period for the specified type will be used. Defaults to a per object value if not specified. zero means delete immediately. */
+  gracePeriodSeconds?: number;
+  /** if set to true, it will trigger an unsafe deletion of the resource in case the normal deletion flow fails with a corrupt object error. A resource is considered corrupt if it can not be retrieved from the underlying storage successfully because of a) its data can not be transformed e.g. decryption failure, or b) it fails to decode into an object. NOTE: unsafe deletion ignores finalizer constraints, skips precondition checks, and removes the object from the storage. WARNING: This may potentially break the cluster if the workload associated with the resource being unsafe-deleted relies on normal deletion flow. Use only if you REALLY know what you are doing. The default value is false, and the user must opt in to enable it */
+  ignoreStoreReadErrorWithClusterBreakingPotential?: boolean;
+  /** A selector to restrict the list of returned objects by their labels. Defaults to everything. */
+  labelSelector?: string;
+  /** limit is a maximum number of responses to return for a list call. If more items exist, the server will set the `continue` field on the list metadata to a value that can be used with the same initial query to retrieve the next set of results. Setting a limit may return fewer than the requested amount of items (up to zero items) in the event all requested objects are filtered out and clients should only use the presence of the continue field to determine whether more results are available. Servers may choose not to support the limit argument and will return all of the available results. If limit is specified and the continue field is empty, clients may assume that no more results are available. This field is not supported if watch is true.
+    
+    The server guarantees that the objects returned when using continue will be identical to issuing a single list call without a limit - that is, no objects created, modified, or deleted after the first request is issued will be included in any subsequent continued requests. This is sometimes referred to as a consistent snapshot, and ensures that a client that is using limit to receive smaller chunks of a very large result can ensure they see all possible objects. If objects are updated during a chunked list the version of the object that was present at the time the first list result was calculated is returned. */
+  limit?: number;
+  /** Deprecated: please use the PropagationPolicy, this field will be deprecated in 1.7. Should the dependent objects be orphaned. If true/false, the "orphan" finalizer will be added to/removed from the object's finalizers list. Either this field or PropagationPolicy may be set, but not both. */
+  orphanDependents?: boolean;
+  /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
+  propagationPolicy?: string;
+  /** resourceVersion sets a constraint on what resource versions a request may be served from. See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersion?: string;
+  /** resourceVersionMatch determines how resourceVersion is applied to list calls. It is highly recommended that resourceVersionMatch be set for list calls where resourceVersion is set See https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions for details.
+    
+    Defaults to unset */
+  resourceVersionMatch?: string;
+  /** `sendInitialEvents=true` may be set together with `watch=true`. In that case, the watch stream will begin with synthetic events to produce the current state of objects in the collection. Once all such events have been sent, a synthetic "Bookmark" event  will be sent. The bookmark will report the ResourceVersion (RV) corresponding to the set of objects, and be marked with `"k8s.io/initial-events-end": "true"` annotation. Afterwards, the watch stream will proceed as usual, sending watch events corresponding to changes (subsequent to the RV) to objects watched.
+    
+    When `sendInitialEvents` option is set, we require `resourceVersionMatch` option to also be set. The semantic of the watch request is as following: - `resourceVersionMatch` = NotOlderThan
+      is interpreted as "data at least as new as the provided `resourceVersion`"
+      and the bookmark event is send when the state is synced
+      to a `resourceVersion` at least as fresh as the one provided by the ListOptions.
+      If `resourceVersion` is unset, this is interpreted as "consistent read" and the
+      bookmark event is send when the state is synced at least to the moment
+      when request started being processed.
+    - `resourceVersionMatch` set to any other value or unset
+      Invalid error is returned.
+    
+    Defaults to true if `resourceVersion=""` or `resourceVersion="0"` (for backward compatibility reasons) and to false otherwise. */
+  sendInitialEvents?: boolean;
+  /** Timeout for the list/watch call. This limits the duration of the call, regardless of any activity or inactivity. */
+  timeoutSeconds?: number;
+};
+export type GetTimeIntervalApiResponse = /** status 200 OK */ TimeInterval;
+export type GetTimeIntervalApiArg = {
+  /** name of the TimeInterval */
+  name: string;
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+};
+export type ReplaceTimeIntervalApiResponse = /** status 200 OK */ TimeInterval | /** status 201 Created */ TimeInterval;
+export type ReplaceTimeIntervalApiArg = {
+  /** name of the TimeInterval */
+  name: string;
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** fieldManager is a name associated with the actor or entity that is making these changes. The value must be less than or 128 characters long, and only contain printable characters, as defined by https://golang.org/pkg/unicode/#IsPrint. */
+  fieldManager?: string;
+  /** fieldValidation instructs the server on how to handle objects in the request (POST/PUT/PATCH) containing unknown or duplicate fields. Valid values are: - Ignore: This will ignore any unknown fields that are silently dropped from the object, and will ignore all but the last duplicate field that the decoder encounters. This is the default behavior prior to v1.23. - Warn: This will send a warning via the standard warning response header for each unknown field that is dropped from the object, and for each duplicate field that is encountered. The request will still succeed if there are no other errors, and will only persist the last of any duplicate fields. This is the default in v1.23+ - Strict: This will fail the request with a BadRequest error if any unknown fields would be dropped from the object, or if any duplicate fields are present. The error returned from the server will contain all unknown and duplicate fields encountered. */
+  fieldValidation?: string;
+  timeInterval: TimeInterval;
+};
+export type DeleteTimeIntervalApiResponse = /** status 200 OK */ Status | /** status 202 Accepted */ Status;
+export type DeleteTimeIntervalApiArg = {
+  /** name of the TimeInterval */
+  name: string;
+  /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
+  pretty?: string;
+  /** When present, indicates that modifications should not be persisted. An invalid or unrecognized dryRun directive will result in an error response and no further processing of the request. Valid values are: - All: all dry run stages will be processed */
+  dryRun?: string;
+  /** The duration in seconds before the object should be deleted. Value must be non-negative integer. The value zero indicates delete immediately. If this value is nil, the default grace period for the specified type will be used. Defaults to a per object value if not specified. zero means delete immediately. */
+  gracePeriodSeconds?: number;
+  /** if set to true, it will trigger an unsafe deletion of the resource in case the normal deletion flow fails with a corrupt object error. A resource is considered corrupt if it can not be retrieved from the underlying storage successfully because of a) its data can not be transformed e.g. decryption failure, or b) it fails to decode into an object. NOTE: unsafe deletion ignores finalizer constraints, skips precondition checks, and removes the object from the storage. WARNING: This may potentially break the cluster if the workload associated with the resource being unsafe-deleted relies on normal deletion flow. Use only if you REALLY know what you are doing. The default value is false, and the user must opt in to enable it */
+  ignoreStoreReadErrorWithClusterBreakingPotential?: boolean;
+  /** Deprecated: please use the PropagationPolicy, this field will be deprecated in 1.7. Should the dependent objects be orphaned. If true/false, the "orphan" finalizer will be added to/removed from the object's finalizers list. Either this field or PropagationPolicy may be set, but not both. */
+  orphanDependents?: boolean;
+  /** Whether and how garbage collection will be performed. Either this field or OrphanDependents may be set, but not both. The default policy is decided by the existing finalizer set in the metadata.finalizers and the resource-specific default policy. Acceptable values are: 'Orphan' - orphan the dependents; 'Background' - allow the garbage collector to delete the dependents in the background; 'Foreground' - a cascading policy that deletes all dependents in the foreground. */
+  propagationPolicy?: string;
+};
+export type UpdateTimeIntervalApiResponse = /** status 200 OK */ TimeInterval | /** status 201 Created */ TimeInterval;
+export type UpdateTimeIntervalApiArg = {
+  /** name of the TimeInterval */
   name: string;
   /** If 'true', then the output is pretty printed. Defaults to 'false' unless the user-agent indicates a browser or command-line HTTP tool (curl and wget). */
   pretty?: string;
@@ -842,101 +1239,31 @@ export type ObjectMeta = {
     Populated by the system. Read-only. More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids */
   uid?: string;
 };
-export type AlertRuleTemplateString = string;
-export type AlertRuleDatasourceUid = string;
-export type AlertRulePromDurationWMillis = string;
-export type AlertRuleRelativeTimeRange = {
-  from: AlertRulePromDurationWMillis;
-  to: AlertRulePromDurationWMillis;
+export type ReceiverIntegration = {
+  disableResolveMessage?: boolean;
+  secureFields?: {
+    [key: string]: boolean;
+  };
+  settings: {
+    [key: string]: {
+      [key: string]: any;
+    };
+  };
+  type: string;
+  uid?: string;
+  version: string;
 };
-export type AlertRuleExpression = {
-  datasourceUID?: AlertRuleDatasourceUid;
-  model: {
-    [key: string]: any;
-  };
-  /** The type of query if this is a query expression */
-  queryType?: string;
-  relativeTimeRange?: AlertRuleRelativeTimeRange;
-  /** Used to mark the expression to be used as the final source for the rule evaluation
-    Only one expression in a rule can be marked as the source
-    For AlertRules, this is the expression that will be evaluated against the alerting condition
-    For RecordingRules, this is the expression that will be recorded */
-  source?: boolean;
-};
-export type AlertRuleExpressionMap = {
-  [key: string]: AlertRuleExpression;
-};
-export type AlertRuleTimeIntervalRef = string;
-export type AlertRulePromDuration = string;
-export type AlertRuleIntervalTrigger = {
-  interval: AlertRulePromDuration;
-};
-export type AlertRuleSpec = {
-  annotations?: {
-    [key: string]: AlertRuleTemplateString;
-  };
-  execErrState: string;
-  expressions: AlertRuleExpressionMap;
-  for?: string;
-  keepFiringFor?: string;
-  labels?: {
-    [key: string]: AlertRuleTemplateString;
-  };
-  missingSeriesEvalsToResolve?: number;
-  noDataState: string;
-  notificationSettings?: {
-    activeTimeIntervals?: AlertRuleTimeIntervalRef[];
-    groupBy?: string[];
-    groupInterval?: AlertRulePromDuration;
-    groupWait?: AlertRulePromDuration;
-    muteTimeIntervals?: AlertRuleTimeIntervalRef[];
-    receiver: string;
-    repeatInterval?: AlertRulePromDuration;
-  };
-  panelRef?: {
-    dashboardUID: string;
-    panelID: number;
-  };
-  paused?: boolean;
+export type ReceiverSpec = {
+  integrations: ReceiverIntegration[];
   title: string;
-  trigger: AlertRuleIntervalTrigger;
 };
-export type AlertRuleOperatorState = {
-  /** descriptiveState is an optional more descriptive state field which has no requirements on format */
-  descriptiveState?: string;
-  /** details contains any extra information that is operator-specific */
-  details?: {
-    [key: string]: {
-      [key: string]: any;
-    };
-  };
-  /** lastEvaluation is the ResourceVersion last evaluated */
-  lastEvaluation: string;
-  /** state describes the state of the lastEvaluation.
-    It is limited to three possible states for machine evaluation. */
-  state: 'success' | 'in_progress' | 'failed';
-};
-export type AlertRuleStatus = {
-  /** additionalFields is reserved for future use */
-  additionalFields?: {
-    [key: string]: {
-      [key: string]: any;
-    };
-  };
-  /** operatorStates is a map of operator ID to operator state evaluations.
-    Any operator which consumes this kind SHOULD add its state evaluation information to this field. */
-  operatorStates?: {
-    [key: string]: AlertRuleOperatorState;
-  };
-};
-export type AlertRule = {
+export type Receiver = {
   /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
   apiVersion: string;
   /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
   kind: string;
   metadata: ObjectMeta;
-  spec: AlertRuleSpec;
-  status?: AlertRuleStatus;
+  spec: ReceiverSpec;
 };
 export type ListMeta = {
   /** continue may be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message. */
@@ -948,10 +1275,10 @@ export type ListMeta = {
   /** Deprecated: selfLink is a legacy read-only field that is no longer populated by the system. */
   selfLink?: string;
 };
-export type AlertRuleList = {
+export type ReceiverList = {
   /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
   apiVersion?: string;
-  items: AlertRule[];
+  items: Receiver[];
   /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
   kind?: string;
   metadata: ListMeta;
@@ -1001,87 +1328,141 @@ export type Status = {
   status?: string;
 };
 export type Patch = object;
-export type RecordingRuleDatasourceUid = string;
-export type RecordingRulePromDurationWMillis = string;
-export type RecordingRuleRelativeTimeRange = {
-  from: RecordingRulePromDurationWMillis;
-  to: RecordingRulePromDurationWMillis;
+export type RoutingTreeRouteDefaults = {
+  group_by?: string[];
+  group_interval?: string;
+  group_wait?: string;
+  receiver: string;
+  repeat_interval?: string;
 };
-export type RecordingRuleExpression = {
-  datasourceUID?: RecordingRuleDatasourceUid;
-  model: {
-    [key: string]: any;
-  };
-  /** The type of query if this is a query expression */
-  queryType?: string;
-  relativeTimeRange?: RecordingRuleRelativeTimeRange;
-  /** Used to mark the expression to be used as the final source for the rule evaluation
-    Only one expression in a rule can be marked as the source
-    For AlertRules, this is the expression that will be evaluated against the alerting condition
-    For RecordingRules, this is the expression that will be recorded */
-  source?: boolean;
+export type RoutingTreeMatcher = {
+  label: string;
+  type: '=' | '!=' | '=~' | '!~';
+  value: string;
 };
-export type RecordingRuleExpressionMap = {
-  [key: string]: RecordingRuleExpression;
+export type RoutingTreeRoute = {
+  active_time_intervals?: string[];
+  continue: boolean;
+  group_by?: string[];
+  group_interval?: string;
+  group_wait?: string;
+  matchers?: RoutingTreeMatcher[];
+  mute_time_intervals?: string[];
+  receiver?: string;
+  repeat_interval?: string;
+  routes?: RoutingTreeRoute[];
 };
-export type RecordingRuleTemplateString = string;
-export type RecordingRulePromDuration = string;
-export type RecordingRuleIntervalTrigger = {
-  interval: RecordingRulePromDuration;
+export type RoutingTreeSpec = {
+  defaults: RoutingTreeRouteDefaults;
+  routes: RoutingTreeRoute[];
 };
-export type RecordingRuleSpec = {
-  expressions: RecordingRuleExpressionMap;
-  labels?: {
-    [key: string]: RecordingRuleTemplateString;
-  };
-  metric: string;
-  paused?: boolean;
-  targetDatasourceUID: string;
-  title: string;
-  trigger: RecordingRuleIntervalTrigger;
-};
-export type RecordingRuleOperatorState = {
-  /** descriptiveState is an optional more descriptive state field which has no requirements on format */
-  descriptiveState?: string;
-  /** details contains any extra information that is operator-specific */
-  details?: {
-    [key: string]: {
-      [key: string]: any;
-    };
-  };
-  /** lastEvaluation is the ResourceVersion last evaluated */
-  lastEvaluation: string;
-  /** state describes the state of the lastEvaluation.
-    It is limited to three possible states for machine evaluation. */
-  state: 'success' | 'in_progress' | 'failed';
-};
-export type RecordingRuleStatus = {
-  /** additionalFields is reserved for future use */
-  additionalFields?: {
-    [key: string]: {
-      [key: string]: any;
-    };
-  };
-  /** operatorStates is a map of operator ID to operator state evaluations.
-    Any operator which consumes this kind SHOULD add its state evaluation information to this field. */
-  operatorStates?: {
-    [key: string]: RecordingRuleOperatorState;
-  };
-};
-export type RecordingRule = {
+export type RoutingTree = {
   /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
   apiVersion: string;
   /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
   kind: string;
   metadata: ObjectMeta;
-  spec: RecordingRuleSpec;
-  status?: RecordingRuleStatus;
+  spec: RoutingTreeSpec;
 };
-export type RecordingRuleList = {
+export type RoutingTreeList = {
   /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
   apiVersion?: string;
-  items: RecordingRule[];
+  items: RoutingTree[];
   /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
   kind?: string;
   metadata: ListMeta;
 };
+export type TemplateGroupTemplateKind = 'grafana' | 'mimir';
+export type TemplateGroupSpec = {
+  content: string;
+  kind: TemplateGroupTemplateKind;
+  title: string;
+};
+export type TemplateGroup = {
+  /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
+  apiVersion: string;
+  /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
+  kind: string;
+  metadata: ObjectMeta;
+  spec: TemplateGroupSpec;
+};
+export type TemplateGroupList = {
+  /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
+  apiVersion?: string;
+  items: TemplateGroup[];
+  /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
+  kind?: string;
+  metadata: ListMeta;
+};
+export type TimeIntervalTimeRange = {
+  end_time: string;
+  start_time: string;
+};
+export type TimeIntervalInterval = {
+  days_of_month?: string[];
+  location?: string;
+  months?: string[];
+  times?: TimeIntervalTimeRange[];
+  weekdays?: string[];
+  years?: string[];
+};
+export type TimeIntervalSpec = {
+  name: string;
+  time_intervals: TimeIntervalInterval[];
+};
+export type TimeInterval = {
+  /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
+  apiVersion: string;
+  /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
+  kind: string;
+  metadata: ObjectMeta;
+  spec: TimeIntervalSpec;
+};
+export type TimeIntervalList = {
+  /** APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources */
+  apiVersion?: string;
+  items: TimeInterval[];
+  /** Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds */
+  kind?: string;
+  metadata: ListMeta;
+};
+export const {
+  useGetApiResourcesQuery,
+  useLazyGetApiResourcesQuery,
+  useListReceiverQuery,
+  useLazyListReceiverQuery,
+  useCreateReceiverMutation,
+  useDeletecollectionReceiverMutation,
+  useGetReceiverQuery,
+  useLazyGetReceiverQuery,
+  useReplaceReceiverMutation,
+  useDeleteReceiverMutation,
+  useUpdateReceiverMutation,
+  useListRoutingTreeQuery,
+  useLazyListRoutingTreeQuery,
+  useCreateRoutingTreeMutation,
+  useDeletecollectionRoutingTreeMutation,
+  useGetRoutingTreeQuery,
+  useLazyGetRoutingTreeQuery,
+  useReplaceRoutingTreeMutation,
+  useDeleteRoutingTreeMutation,
+  useUpdateRoutingTreeMutation,
+  useListTemplateGroupQuery,
+  useLazyListTemplateGroupQuery,
+  useCreateTemplateGroupMutation,
+  useDeletecollectionTemplateGroupMutation,
+  useGetTemplateGroupQuery,
+  useLazyGetTemplateGroupQuery,
+  useReplaceTemplateGroupMutation,
+  useDeleteTemplateGroupMutation,
+  useUpdateTemplateGroupMutation,
+  useListTimeIntervalQuery,
+  useLazyListTimeIntervalQuery,
+  useCreateTimeIntervalMutation,
+  useDeletecollectionTimeIntervalMutation,
+  useGetTimeIntervalQuery,
+  useLazyGetTimeIntervalQuery,
+  useReplaceTimeIntervalMutation,
+  useDeleteTimeIntervalMutation,
+  useUpdateTimeIntervalMutation,
+} = injectedRtkApi;
